requirements.txt
absl-py==1.2.0
astor==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1660605382950/work
astunparse==1.6.3
attrdict==2.0.1
Babel==2.10.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bce-python-sdk==0.8.74
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
cssselect==1.1.0
cssutils==2.5.1
cycler==0.11.0
Cython==0.29.32
debugpy @ file:///D:/bld/debugpy_1660619093354/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
et-xmlfile==1.1.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1660472483358/work
Flask==2.2.2
Flask-Babel==2.0.0
flask-ngrok==0.0.25
flatbuffers==1.12
fonttools==4.35.0
future==0.18.2
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.21.1
imgaug==0.4.0
importlib-metadata==4.12.0
ipykernel @ file:///D:/bld/ipykernel_1657295151080/work
ipython @ file:///D:/bld/ipython_1653755022901/work
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1659959867326/work
Jinja2==3.1.2
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///D:/bld/jupyter_core_1658332534377/work
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lmdb==1.3.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1648959695634/work
networkx==2.8.5
numpy==1.23.2
oauthlib==3.2.0
opencv-contrib-python==4.4.0.46
opencv-python==4.6.0.66
openpyxl==3.0.10
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
paddle-bfloat==0.1.7
paddleocr==2.5.0.3
paddlepaddle==2.3.2
paddlepaddle-gpu==2.3.1
pandas==1.4.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.2.0
poppler-utils==0.1.0
premailer==3.10.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1656332401605/work
protobuf==3.19.4
psutil @ file:///D:/bld/psutil_1653089356250/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pycryptodome==3.15.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1650904496387/work
pyngrok==4.1.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
pytesseract==0.3.9
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-Levenshtein==0.12.2
pytz==2022.2.1
PyWavelets==1.3.0
pywin32==303
PyYAML==6.0
pyzmq @ file:///D:/bld/pyzmq_1660329149353/work
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.0
Shapely==1.8.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1660472507741/work
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tesseract==0.1.3
tifffile==2022.8.12
tornado @ file:///D:/bld/tornado_1656937958156/work
tqdm==4.64.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1655411388954/work
typing_extensions==4.3.0
urllib3==1.26.11
visualdl==2.3.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
